build: consolidate temporal logic library, examples and dafny reference
