+++ src/temporal-logic-lib/examples/abc_state_machine/liveness.rs
@@ -1,7 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
-use crate::simple_example::state_machine::*;
+use crate::examples::abc_state_machine::state_machine::*;
 use crate::temporal_logic::*;
 use builtin::*;
 use builtin_macros::*;
