+++ src/temporal-logic-lib/examples/abc_state_machine/state_machine.rs
