+++ src/temporal-logic-lib/main.rs
@@ -1,10 +1,8 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 
-pub mod concurrent_controller_example;
-pub mod controller_example;
+pub mod examples;
 pub mod pervasive;
-pub mod simple_example;
 pub mod temporal_logic;
 
 fn main() {}
