+++ src/temporal-logic-lib/temporal_logic.rs
